Warning cleanup.
- Changed remaining printf to iostreams, again for int width related
  warnings.
- Fixed documentation (Shift + F8 and F8 were mixed up). Also, made
  the SAVE documentation clearer (I think)...

+++ source/src/Screen.cc
@@ -217,8 +217,8 @@ void Screen::update()
     if (tape.pulseData.size())
     {
         char str[64];
-        size_t percent = 100 * tape.pointer / tape.pulseData.size();
-        snprintf(str, 64, "SpecIde [%03zu%%]", percent);
+        unsigned int percent = 100 * tape.pointer / tape.pulseData.size();
+        snprintf(str, 64, "SpecIde [%03u%%]", percent);
         window.setTitle(str);
     }
